Folder system organization and restructure
